Added cloak description
The second part is to sort of explain why it doesn't count as clothing. Also, why do all the bags/quivers count as clothing for nudity? The CE backpack doesn't count as clothing, and neither do other similar items like the Sash. If anything, I'd say this could count as clothing, not the bags. But also please don't make this count as clothing, i like my nudists not dying of frostbite.